+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "TimexCharts",
-  "version": "0.1.3",
+  "name": "timexcharts",
+  "version": "0.1.4",
   "description": "A library for creating Timex-designed Charts using Chart.js",
   "main": "timexcharts.js",
   "scripts": {
