--- step_count_reference.html
@@ -1,91 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>Timex Charts Demo</title>
-    <style>
-      {
-        margin: 0;
-        padding: 0;
-        font-family: sans-serif;
-      }
-      .chartMenu {
-        width: 100vw;
-        height: 40px;
-        background: #1A1A1A;
-        color: rgba(75, 192, 192, 1);
-      }
-      .chartMenu p {
-        padding: 10px;
-        font-size: 20px;
-      }
-      .chartCard {
-        width: 100vw;
-        height: calc(100vh - 40px);
-        background: rgba(75, 192, 192, 0.2);
-        display: flex;
-        align-items: center;
-        justify-content: center;
-      }
-      .chartBox {
-        width: 700px;
-        padding: 20px;
-        border-radius: 20px;
-        border: solid 3px rgba(75, 192, 192, 1);
-        background: white;
-      }
-    </style>
-  </head>
-  <body>
-    <div class="chartMenu">
-      <p>Step Count Bar Chart</p>
-    </div>
-    <div class="chartCard">
-      <div class="chartBox">
-        <canvas id="myChart"></canvas>
-      </div>
-    </div>
-  
-    <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
-    <script src="https://cdn.jsdelivr.net/gh/timexgroup/TimexCharts_Web_JS@v0.1.1/timexcharts.js"></script>
-
-    <script>
-    let randomNumbers = [];
-    let labels = [];
-
-    // Generate 96 random numbers and push them into the array
-    for (let i = 0; i < 24; i++) {
-      randomNumbers.push(Math.random());
-      labels.push(String(i))
-    }
-
-    // data block
-    const data = {
-      // labels: ['12:00 AM', 'Blue', 'Yellow', '12:00 PM', 'Purple', 'Orange'],
-      labels: labels,
-      datasets: [{
-        data: randomNumbers,
-        // data: [12, 19, 3, 5, 2, 3],
-        backgroundColor: 'rgba(62, 177, 200, 1.0)',
-        borderColor: 'rgba(62, 177, 200, 1.0)',
-        borderWidth: 1,
-        barPercentage: 1,
-        categoryPercentage: 1,        
-      }]
-    };
-
-    // init render block
-    const myChart = new TimexCharts.StepCountBarChart(
-      document.getElementById('myChart'),
-      data
-    );
-    </script>
-
-
-  </body>
-</html>
-
-
-
-
